requirements.txt
# Core dependencies
google-genai>=1.0.0
gradio>=4.0.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0

# Backend - FastAPI
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database - Supabase
supabase>=2.0.0

# Storage - Cloudflare R2 (S3-compatible)
boto3>=1.34.0

# Auth
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
